Улучшен репозиторий, добавлена возможность выбрать город

+++ vacancy-tracker-bot/app/src/main/java/vacancy_tracker/model/api/dto/VacancySearchFilter.java
@@ -18,7 +18,7 @@ public class VacancySearchFilter {
     private Integer minSalary;
     private Integer maxSalary;
     private Location location;
-    private Integer experienceFrom;
+    private Float experienceFrom;
 
     private String modifiedFrom;
 }
